+++ PRIO/src/Modules/Hierarchy/Installations/ViewModels/UpdateInstallationViewModel.cs
@@ -15,7 +15,7 @@ namespace PRIO.src.Modules.Hierarchy.Installations.ViewModels
         [StringLength(60, ErrorMessage = "CodInstallationAnp cannot exceed 60 characters.")]
         public string? CodInstallationAnp { get; set; }
         [DecimalPrecision(12)]
-        public double? GasSafetyBurnVolume { get; set; }
+        public decimal? GasSafetyBurnVolume { get; set; }
         public string? Description { get; set; }
     }
 }
